More fine-grained user permissions.
Some permissions were added so there's a small DB change (adding the 
Permission objects) which you can see in the models.py

To change Anonymous user access permissions, I use an ANONYMOUS_USER 
named user; its kind of hacky, but using a group was not going to work 
out. And I store the string in settings. Which makes it slightly more 
acceptable.

The system uses a few decorators rather than just one, for responses 
that want json/xml errors, html error pages, or a redirect to login.

A ton of things require permissions now, so it might break shit. check 
grep @permission to get an idea of it.