Adding two template tags 'set' and 'unset'. This allows arbitrary variable setting in templates.
The use of this for now is to eliminate redundancy in songlist base. It basically
allows songlist_song_row to take arguments. Now if you set 'current' to something
when you branch to songlist_song_row, then it will render the song as the 'current'
song. I also use set to assign song to the value of current_song. After the include,
I unset them. 

This style tag is very useful, and I don't understand why Django doesn't come with them, 
or at least some way to provide additional context to subrendered templates. That style
thing is common in other template systems.