Don't use eventlet for scraping.
It's not reliable enough without some debugging.  We get HTTP timeouts when
downloading movie covers, which results in a cover image that is actually an
HTML error document.  This causes PIL to be unable to parse the file as an
image, which manifests as exceptions about subscripting the None type when
Django tries to find the width of the image.